[feature] exports rule-test-harness to be used with other frameworks and in-repo lint rules
